+++ src/main/java/com/rinha/frauddetector/adapter/loader/ReferenceItem.java
@@ -1,3 +1,3 @@
 package com.rinha.frauddetector.adapter.loader;
 
-public record ReferenceItem(double[] vector, String label) {}
+public record ReferenceItem(float[] vector, String label) {}
