+++ src/main/java/com/rinha/frauddetector/domain/TransactionVector.java
@@ -17,10 +17,10 @@ public record TransactionVector(float[] features) {
     }
   }
 
-  public double distanceTo(TransactionVector other) {
-    double sum = 0;
+  public float distanceTo(TransactionVector other) {
+    float sum = 0;
     for (int i = 0; i < VECTOR_SIZE; i++) {
-      double diff = this.features[i] - other.features[i];
+      float diff = this.features[i] - other.features[i];
       sum += diff * diff;
     }
     return sum;
@@ -29,7 +29,7 @@ public record TransactionVector(float[] features) {
   public static TransactionVector fromRequest(
       FraudRequest request,
       NormalizationConstants constants,
-      Map<String, Double> mccRiskMap) {
+      Map<String, Float> mccRiskMap) {
 
     final var transaction = request.transaction();
     final var customer = request.customer();
@@ -41,11 +41,11 @@ public record TransactionVector(float[] features) {
     final int hour = instant.atZone(ZoneOffset.UTC).getHour();
     final int dayOfWeek = instant.atZone(ZoneOffset.UTC).getDayOfWeek().getValue() - 1;
 
-    final double amountRatio = customer.avg_amount() > 0
-        ? transaction.amount() / customer.avg_amount()
-        : 0.0;
+    final float amountRatio = customer.avg_amount() > 0
+        ? (float) (transaction.amount() / customer.avg_amount())
+        : 0.0f;
 
-    double minutesSinceLastTx = 0;
+    float minutesSinceLastTx = 0;
     if (lastTransaction != null) {
       final Instant lastInstant = Instant.parse(lastTransaction.timestamp());
       minutesSinceLastTx = Duration.between(lastInstant, instant).toMinutes();
@@ -54,13 +54,13 @@ public record TransactionVector(float[] features) {
     float[] v = new float[VECTOR_SIZE];
 
     // 0: amount
-    v[0] = clamp((float) (transaction.amount() / constants.max_amount()));
+    v[0] = clamp(transaction.amount() / constants.max_amount());
 
     // 1: installments
     v[1] = clamp(transaction.installments() / (float) constants.max_installments());
 
     // 2: amount_vs_avg
-    v[2] = clamp((float) (amountRatio / constants.amount_vs_avg_ratio()));
+    v[2] = clamp(amountRatio / constants.amount_vs_avg_ratio());
 
     // 3: hour_of_day
     v[3] = clamp(hour / 23.0f);
@@ -70,16 +70,16 @@ public record TransactionVector(float[] features) {
 
     // 5: minutes_since_last_tx
     v[5] = lastTransaction != null
-        ? clamp((float) (minutesSinceLastTx / constants.max_minutes()))
+        ? clamp(minutesSinceLastTx / constants.max_minutes())
         : -1.0f;
 
     // 6: km_from_last_tx
     v[6] = lastTransaction != null
-        ? clamp((float) (lastTransaction.km_from_current() / constants.max_km()))
+        ? clamp(lastTransaction.km_from_current() / constants.max_km())
         : -1.0f;
 
     // 7: km_from_home
-    v[7] = clamp((float) (terminal.km_from_home() / constants.max_km()));
+    v[7] = clamp(terminal.km_from_home() / constants.max_km());
 
     // 8: tx_count_24h
     v[8] = clamp(customer.tx_count_24h() / (float) constants.max_tx_count_24h());
@@ -94,10 +94,10 @@ public record TransactionVector(float[] features) {
     v[11] = customer.known_merchants().contains(merchant.id()) ? 0.0f : 1.0f;
 
     // 12: mcc_risk
-    v[12] = mccRiskMap.getOrDefault(merchant.mcc(), 0.5).floatValue();
+    v[12] = mccRiskMap.getOrDefault(merchant.mcc(), (float) 0.5);
 
     // 13: merchant_avg_amount
-    v[13] = clamp((float) (merchant.avg_amount() / constants.max_merchant_avg_amount()));
+    v[13] = clamp(merchant.avg_amount() / constants.max_merchant_avg_amount());
 
     return new TransactionVector(v);
   }
