+++ src/main/java/com/rinha/frauddetector/dto/CustomerDTO.java
@@ -3,7 +3,7 @@ package com.rinha.frauddetector.dto;
 import java.util.List;
 
 public record CustomerDTO(
-    double avg_amount,
+    float avg_amount,
     int tx_count_24h,
     List<String> known_merchants
 ) {}
