+++ src/main/java/com/rinha/frauddetector/dto/FraudResponse.java
@@ -2,5 +2,5 @@ package com.rinha.frauddetector.dto;
 
 public record FraudResponse(
     boolean approved,
-    double fraud_score
+    float fraud_score
 ) {}
