+++ src/main/java/com/rinha/frauddetector/dto/MerchantDTO.java
@@ -3,5 +3,5 @@ package com.rinha.frauddetector.dto;
 public record MerchantDTO(
     String id,
     String mcc,
-    double avg_amount
+    float avg_amount
 ) {}
