+++ src/main/java/com/rinha/frauddetector/dto/TerminalDTO.java
@@ -3,5 +3,5 @@ package com.rinha.frauddetector.dto;
 public record TerminalDTO(
     boolean is_online,
     boolean card_present,
-    double km_from_home
+    float km_from_home
 ) {}
